requirements.txt
alembic==1.10.3
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.24.2
pandas==2.0.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
